move Cache into `io` folder
